make session inherit from session_handle and remove wrapper functions
